Add calculate_checksum methods to IP and TCP
- The normal library behavior is to get or set the checksum in the packet
  using checksum() getter/setter methods. The correct checksum is only
  calculated, and automatically set, just prior to serializing the data.
  There is no way to simply check if the received checksum is valid.

- This commit adds calculate_checksum methods to IP and TCP which
  calculate the correct checksum without modifying the packet. The library
  user may compare this checksum against the checksum stored in the packet
  by calling checksum().

- tcp_test.cpp currently fails 1 test.

+++ include/tins/utils/checksum_utils.h
@@ -50,7 +50,7 @@ namespace Utils {
  * \return Returns the checksum between start and end (non inclusive) 
  * in network endian
  */
-TINS_API uint32_t do_checksum(const uint8_t* start, const uint8_t* end);
+TINS_API uint16_t do_checksum(const uint8_t* start, const uint8_t* end);
 
 /** 
  * \brief Computes the 16 bit sum of the input buffer.
